* tldr.el: alter some strings and docstrings.
The Emacs manual recommends that documentation strings are complete
sentences, beginning with a capital letter and ending with a full
stop, so I have made sure that all of them fit this.

I also reworded some of the sentences so that they sound more natural.
Nothing that I did changed the behaviour of the package, it was all
simply quality of life.